refactor(dm): split GroupConversationScreen into cohesive modules (#703)
GroupConversationScreen.tsx was at the 1000-line cap (#703 baseline 1015).
Mirror the ConversationScreen split: separate by responsibility so it drops to
731 lines, off the baseline.

- styles/GroupConversationScreen.styles.ts — presentation (createGroupConversationScreenStyles)
- hooks/useGroupComposerActions.ts        — send/upload/share orchestration +
                                            in-flight flags (sibling of
                                            useConversationComposerActions),
                                            funnelling through sendText
- screens/GroupConversationScreen.tsx     — now layout + wiring + renderMessage

Behaviour unchanged (verbatim moves, strongly-typed wiring). Verified on the LP
emulator: opened the group thread — header (name / 3 members / edit / leave),
message history, a shared-contact card, the composer, and the attach panel
(Camera/Gallery/GIF/Location/Zap/Invoice/Contact) all render and respond.
Removes the GroupConversationScreen baseline from check-file-size.sh.
